Reference

Contents

Reference

Contents

Index

Index

OhMyPlots.OhMyPlots

OhMyPlots.linestyle_palette

OhMyPlots.CurveVar

OhMyPlots.PlotAttributsXYLine

OhMyPlots.PlotAttributsXYLine

OhMyPlots.PlotDataXYLine

OhMyPlots.Var

OhMyPlots.XY

OhMyPlots.XY_Line_Scatter_Theme

OhMyPlots.XY_Theme

OhMyPlots.__check_number_of_x_vars

OhMyPlots.__check_number_of_y_vars

OhMyPlots._apply_xlims

OhMyPlots._apply_xlims

OhMyPlots._apply_ylims

OhMyPlots._apply_ylims

OhMyPlots._check_xy_vars

OhMyPlots._create_pallette

OhMyPlots._draw_scatterlines

OhMyPlots._make_fig_ax

OhMyPlots._post_process_figure

OhMyPlots._pre_process_figure

OhMyPlots.ax_barier

OhMyPlots.xy_label!

OhMyPlots.OhMyPlots — Module

OhMyPlots

This module provides a set of tools to create beautiful plots using Makie.jl.

OhMyPlots.linestyle_palette — Constant

pallete

OhMyPlots.CurveVar — Type

struct CurveVar{S<:AbstractString} <: VariableAbstract

Curve variable type representing a curve with a legend.

Fields

legend::S: The legend of the curve.

OhMyPlots.PlotAttributsXYLine — Type

struct PlotAttributsXYLine{S,T,P} <: PlotAttributsAbstract

The `PlotAttributsXYLine` struct represents the attributes of an XY line plot.

# Fields
- `data::Vector{PlotDataXYLine{S,T,P}}`: Vector of XY line plot data.
- `save_path::String`: Path to save the plot.
- `save_format::String`: Format to save the plot.
- `title::Union{LaTeXString,Nothing}`: Title of the plot (optional).
- `is_scatter::Bool`: Flag indicating whether the plot is a scatter plot or not.
- `ylim::Union{Nothing, Tuple{Float64,Float64}}`: Y-axis limits of the plot (optional).
- `xlim::Union{Nothing, Tuple{Float64,Float64}}`: X-axis limits of the plot (optional).
- `n_curve::Int`: Number of curves in the plot.
- `n_each_curve::Int`: Number of data points in each curve.

OhMyPlots.PlotAttributsXYLine — Method

PlotAttributsXYLine(data::Vector{PlotDataXYLine{S,T,P}}, save_path::Q, save_format::Q, n_curve::Int, n_each_curve::Int; kwargs...) where {S,T,P,Q<:AbstractString}

Constructs a `PlotAttributsXYLine` object.

# Arguments
- `data::Vector{PlotDataXYLine{S,T,P}}`: The data for the plot.
- `save_path::Q`: The path to save the plot.
- `save_format::Q`: The format to save the plot.
- `n_curve::Int`: The number of curves in the plot.
- `n_each_curve::Int`: The number of points in each curve.
- `kwargs...`: Additional keyword arguments.

# Returns
A `PlotAttributsXYLine{S,T,P}` object.

OhMyPlots.PlotDataXYLine — Type

struct PlotDataXYLine{S,T,P}<: PlotDataAbstract

The PlotDataXYLine struct represents a type of plot data for XY line plots.

Fields

XVar::Var{S,T}: The variable representing the X-axis data.

YVar::Var{S,P}: The variable representing the Y-axis data.

CurveVars::CurveVar{S}: The variable representing the curve data.

OhMyPlots.Var — Type

struct Var{S<:AbstractString,T<:Union{AbstractFloat,Int, AbstractString}} <: VariableAbstract

The Var struct represents a variable with a name, unit, and a vector of values.

Fields

name::S: The name of the variable.

unit::S: The unit of measurement for the variable.

value::Vector{T}: A vector of values for the variable.

OhMyPlots.XY — Method

XY(Pattr::PlotAttributsXYLine{S,T,P}) where {S,T,P}

This function generates an XY plot based on the given plot attributes.

# Arguments
- `Pattr::PlotAttributsXYLine{S,T,P}`: The plot attributes object containing information such as data, save path, save format, title, scatter option, y-axis limits, x-axis limits, number of curves, and number of points in each curve.

# Returns
- `fig`: The generated figure object.

OhMyPlots.XY_Line_Scatter_Theme — Method

XYLineScatter_Theme(;size=size, fonts=fonts, fontsize=fontsize, font=font, axis=axis, Legend=Legend, Colorbar=Colorbar,cycle= cycle, palette=palette, markersize=markersize, markercolor=markercolor, strokewidth=strokewidth, position=position, nbanks=nbanks,labelsize=labelsize, orientation=orientation)

Create a theme for XY plots with lines and scatter plots

# Arguments
- size: Tuple{Int,Int} = (600, 480)
- fonts: NamedTuple = (;regular = Makie.assetpath("times.ttf"),
                bold_italic  = Makie.assetpath("timesbi.ttf"),
                bold= Makie.assetpath("timesbd.ttf"),
                italic = Makie.assetpath("timesi.ttf")
                )
- fontsize: Int = 20
- font: Symbol = :bold
- axis: NamedTuple = (
        xlabelsize=28,xlabelpadding=+5,
        ylabelsize=28,ylabelpadding=+5,
        xgridstyle=:dash, ygridstyle=:dash,
        xgridwidth = 1.5, ygridwidth = 1.5,
        xtickalign=1, ytickalign=1,
        yticksize=10, xticksize=10,
        markersize=16, markercolor= :white, strokewidth=2.0, 
        )
- Legend: NamedTuple = (framecolor=(:black, 0.5), backgroundcolor=(:white, 0.5),
            position=:lt, nbanks=3,labelsize=12, orientation=:horizontal)
- Colorbar: NamedTuple = (ticksize=16, tickalign=1, spinewidth=0.5)
- cycle: Cycle = Cycle([:color, :linestyle, :marker], covary=true)
- palette: NamedTuple = (linestyle = linestyle_palette,marker = marker_palette,color = color_palette)
- markersize: Int = 8
- markercolor: Symbol = :white
- strokewidth: Float64 = 2.0
- position: Symbol = :lt
- nbanks: Int = 3
- labelsize: Int = 12
- orientation: Symbol = :horizontal

OhMyPlots.XY_Theme — Method

XY_Theme(;size=size, fonts=fonts, fontsize=fontsize, font=font, axis=axis, Legend=Legend, Colorbar=Colorbar,cycle= cycle, palette=palette, markersize=markersize, markercolor=markercolor, strokewidth=strokewidth, position=position, nbanks=nbanks,labelsize=labelsize, orientation=orientation)

Create a theme for XY plots

# Arguments
- size: Tuple{Int,Int} = (600, 480)
- fonts: NamedTuple = (;regular = Makie.assetpath("times.ttf"),
                bold_italic  = Makie.assetpath("timesbi.ttf"),
                bold= Makie.assetpath("timesbd.ttf"),
                italic = Makie.assetpath("timesi.ttf")
                )
- fontsize: Int = 20
- font: Symbol = :bold
- axis: NamedTuple = (
        xlabelsize=28,xlabelpadding=+5,
        ylabelsize=28,ylabelpadding=+5,
        xgridstyle=:dash, ygridstyle=:dash,
        xgridwidth = 1.5, ygridwidth = 1.5,
        xtickalign=1, ytickalign=1,
        yticksize=10, xticksize=10,
        markersize=16, markercolor= :white, strokewidth=2.0, 
        )
- Legend: NamedTuple = (framecolor=(:black, 0.5), backgroundcolor=(:white, 0.5),
            position=:lt, nbanks=3,labelsize=12, orientation=:horizontal)
- Colorbar: NamedTuple = (ticksize=16, tickalign=1, spinewidth=0.5)
- cycle: Cycle = Cycle([:color, :linestyle, :marker], covary=true)
- palette: NamedTuple = (linestyle = linestyle_palette,marker = marker_palette,color = color_palette)
- markersize: Int = 8
- markercolor: Symbol = :white
- strokewidth: Float64 = 2.0
- position: Symbol = :lt
- nbanks: Int = 3
- labelsize: Int = 12
- orientation: Symbol = :horizontal

linestyle_palette = (:solid,(:dot,:dense) ,:dash, :dashdot,:dashdotdot,(:dot,:loose))
marker_palette = [:circle,:rect,:cross,:utriangle, :dtriangle,:xcross]
color_palette = [:red, :blue, :black,:green,:orange,:pink]

OhMyPlots.__check_number_of_x_vars — Method

__check_number_of_x_vars(data::Vector{PlotDataXYLine{S,T,P}}) where {S,T,P}

Check the number of XVars in the given data.

# Arguments
- `data::Vector{PlotDataXYLine{S,T,P}}`: A vector of `PlotDataXYLine` objects.

# Returns
- `nothing`: If the number of XVars is consistent.

# Throws
- `error`: If the number of XVars is not consistent.

OhMyPlots.__check_number_of_y_vars — Method

__check_number_of_y_vars(data::Vector{PlotDataXYLine{S,T,P}}) where {S,T,P}

Check the number of YVars in the given data.

# Arguments
- `data::Vector{PlotDataXYLine{S,T,P}}`: A vector of `PlotDataXYLine` objects.

# Returns
- `nothing`: If the number of YVars is consistent.

# Throws
- `ErrorException`: If the number of YVars is not consistent.

OhMyPlots._apply_xlims — Method

_apply_xlims(ax::Axis, x::Nothing)

Apply the x-axis limits to the given `ax` Axis object.

# Arguments
- `ax::Axis`: The Axis object to apply the limits to.
- `x::Nothing`: The x-axis limits. Since it is `Nothing`, no limits are applied.

# Returns
- `ax::Axis`: The updated Axis object.

OhMyPlots._apply_xlims — Method

_apply_xlims(ax::Axis, x::Tuple{T,T}) where T<:Real

Apply the x-axis limits to the given `ax` Axis object.

# Arguments
- `ax::Axis`: The Axis object to apply the limits to.
- `x::Tuple{T,T}`: The x-axis limits as a tuple of two elements.

# Returns
- `ax::Axis`: The modified Axis object.

OhMyPlots._apply_ylims — Method

_apply_ylims(ax::Axis, x::Nothing)

Apply y-axis limits to the given axis.

Parameters
----------
ax : Axis
    The axis to apply the y-axis limits to.
x : Nothing
    The value representing no y-axis limits.

Returns
-------
Axis
    The modified axis.

OhMyPlots._apply_ylims — Method

_apply_ylims(ax::Axis, x::Tuple{T,T}) where T<:Real

Apply the y-axis limits to the given `ax` Axis object.

# Arguments
- `ax::Axis`: The Axis object to apply the y-axis limits to.
- `x::Tuple{T,T}`: The y-axis limits as a tuple of two elements.

# Returns
- `ax`: The modified Axis object with the y-axis limits applied.

OhMyPlots._check_xy_vars — Method

_check_xy_vars(data::Vector{PlotDataXYLine{S,T,P}}) where {S,T,P}

Check the number of x and y variables in the given vector of `PlotDataXYLine` objects.

# Arguments
- `data`: A vector of `PlotDataXYLine` objects.

OhMyPlots._create_pallette — Method

_create_pallette(data::Vector{PlotDataXYLine{S,T,P}}, n_curve, n_each_curve) where {S,T,P}

Create a palette and color order for plotting multiple curves.

# Arguments
- `data::Vector{PlotDataXYLine{S,T,P}}`: Vector of plot data for XY line plots.
- `n_curve`: Number of curves to plot.
- `n_each_curve`: Number of colors to use for each curve.

# Returns
- `palette`: A tuple containing the linestyle, marker, and color palettes.
- `color_order`: An ordered list of colors to use for each curve.

OhMyPlots._draw_scatterlines — Method

_draw_scatterlines(data::Vector{PlotDataXYLine{S,T,P}}, ax, fig, color_order) where {S,T,P}

Draws scatterlines on the given axes ax using the data provided in data. Each element in data represents a line plot with x-values XVar.value and y-values YVar.value. The label for each line plot is taken from CurveVars.legend. The scatterlines are drawn with the specified color_order for strokecolor and color.

Arguments

data::Vector{PlotDataXYLine{S,T,P}}: Vector of line plot data.

ax: Axes object to draw the scatterlines on.

fig: Figure object to which the axes belong.

color_order: Vector of colors to use for strokecolor and color.

Returns

ax: The modified axes object.

fig: The unmodified figure object.

OhMyPlots._make_fig_ax — Method

_make_fig_ax(cycle, palette, xy_label, title)

Constructs a figure and axis for plotting XY data.

Arguments

cycle::Any: The cycle parameter for ScatterLines.

palette::Any: The palette parameter for the figure.

xy_label::Tuple: A tuple containing the x-axis and y-axis labels and units.

title::Any: The title of the plot.

Returns

ax::Axis: The constructed axis.

fig::Figure: The constructed figure.

OhMyPlots._post_process_figure — Method

_post_process_figure(ax, ylim, xlim)

Post-processes the figure by applying the specified y-limits (`ylim`) and x-limits (`xlim`) to the axes (`ax`).

# Arguments
- `ax`: The axes object to be post-processed.
- `ylim`: The y-limits to be applied to the axes.
- `xlim`: The x-limits to be applied to the axes.

# Returns
The post-processed axes object.

OhMyPlots._pre_process_figure — Method

_pre_process_figure(data, save_path, save_format, n_curve, n_each_curve)

Pre-processes the figure data before saving it.

Arguments

data: The figure data.

save_path: The path where the figure will be saved.

save_format: The format in which the figure will be saved.

n_curve: The number of curves in the figure.

n_each_curve: The number of points in each curve.

Returns

image_path: The path of the saved image.

cycle: The cycle of line styles and markers.

palette: The color palette for the figure.

color_order: The order of colors in the palette.

xy_label: The labels for the x and y axes.

OhMyPlots.ax_barier — Method

ax_barier(ax::Axis)

A function that returns the input `ax` object. It is just a barrier function for type stability.

# Arguments
- `ax::Axis`: The input `Axis` object.

# Returns
- `ax`: The input `Axis` object.

OhMyPlots.xy_label! — Method

xy_label!(data, xy_label)

Set the labels for the x and y variables in the given `data` structure.

# Arguments
- `data`: The data structure containing the x and y variables.
- `xy_label`: An array to store the labels for the x and y variables. The labels are assigned to the elements of `xy_label` in the following order: x variable name, x variable unit, y variable name, y variable unit.